migrate token gc
